+++ changelog.md
@@ -4,9 +4,11 @@
 ### adding
 1. CatSSR! button in homepage
 2. Adding back to homepage option
+3. Adding counter promoting system in CatSSR!
 ### improving
 1. Using new create random number function in Cat!
-2. Using Promote instead of alert in hint 
+2. Using Promote instead of alert in hint
+3. improving image presenting  in CatSSR!
 
 ## update 20201026
 ### adding